requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
oracledb>=2.0.0
python-dotenv>=1.0.0
werkzeug>=3.0.0
waitress>=3.0.0
rich>=13.0.0
